Make rewards a staking submodule
