+++ crml/staking/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2018"
 repository = "https://github.com/cennznet/cennznet"
 description = "CENNZnet staking pallet"
 
-
 [dependencies]
 serde = { version = "1.0.101", optional = true }
 codec = { package = "parity-scale-codec", version = "1.3.5", default-features = false, features = ["derive"] }
@@ -14,6 +13,7 @@ sp-keyring = { optional = true, git = "https://github.com/plugblockchain/plug-bl
 sp-npos-elections = { default-features = false, git = "https://github.com/plugblockchain/plug-blockchain", branch = "spike/2.0" }
 sp-std = { default-features = false, git = "https://github.com/plugblockchain/plug-blockchain", branch = "spike/2.0" }
 sp-io ={ default-features = false, git = "https://github.com/plugblockchain/plug-blockchain", branch = "spike/2.0" }
+sp-arithmetic = { default-features = false, git = "https://github.com/plugblockchain/plug-blockchain", branch = "spike/2.0" }
 sp-runtime = { default-features = false, git = "https://github.com/plugblockchain/plug-blockchain", branch = "spike/2.0" }
 sp-staking = { default-features = false, git = "https://github.com/plugblockchain/plug-blockchain", branch = "spike/2.0" }
 frame-support = { default-features = false, git = "https://github.com/plugblockchain/plug-blockchain", branch = "spike/2.0" }
@@ -36,6 +36,7 @@ default = ["std"]
 std = [
 	"codec/std",
 	"serde",
+	"sp-arithmetic/std",
 	"sp-io/std",
 	"sp-keyring",
 	"sp-std/std",
