+++ runtime/Cargo.toml
@@ -74,7 +74,6 @@ hex-literal = { version = "0.3.1", optional = true }
 cennznet-primitives = { path = "../primitives", default-features = false }
 crml-cennzx = { path = "../crml/cennzx", default-features = false }
 crml-cennzx-rpc-runtime-api = { path = "../crml/cennzx/rpc/runtime-api", default-features = false }
-crml-rewards = { path = "../crml/rewards", default-features = false }
 crml-staking = { path = "../crml/staking", default-features = false}
 crml-staking-reward-curve = { path = "../crml/staking/reward-curve", default-features = false}
 crml-sylo = { path = "../crml/sylo", default-features = false}
@@ -129,7 +128,6 @@ std = [
 	"cennznet-primitives/std",
 	"crml-cennzx/std",
 	"crml-cennzx-rpc-runtime-api/std",
-	"crml-rewards/std",
 	"crml-staking/std",
 	"crml-sylo/std",
 	"crml-transaction-payment/std",
